+++ Tests/Integration/MSFT_Language.config.ps1
@@ -1,35 +1,43 @@
-
 # Integration Test Config Template Version: 1.0.0
 configuration MSFT_Language_Config {
-    Param
+    param
     (
-        [Parameter(Mandatory=$true)]
+        [Parameter(Mandatory = $true)]
         [Int]
         $LocationID,
-        [Parameter(Mandatory=$true)]
+
+        [Parameter(Mandatory = $true)]
         [String]
         $MUILanguage,
-        [Parameter(Mandatory=$true)]
+
+        [Parameter(Mandatory = $true)]
         [String]
         $MUIFallbackLanguage,
-        [Parameter(Mandatory=$true)]
+
+        [Parameter(Mandatory = $true)]
         [String]
         $SystemLocale,
-        [Parameter(Mandatory=$true)]
+
+        [Parameter(Mandatory = $true)]
         [String]
         $AddInputLanguages,
-        [Parameter(Mandatory=$true)]
+
+        [Parameter(Mandatory = $true)]
         [String]
         $RemoveInputLanguages,
-        [Parameter(Mandatory=$true)]
+
+        [Parameter(Mandatory = $true)]
         [String]
         $UserLocale
     )
-    Import-DscResource -ModuleName ComputerManagementDsc
-    node localhost {
 
-        Language Integration_Test {
-            IsSingleInstance = "Yes" 
+    Import-DscResource -ModuleName ComputerManagementDsc
+
+    node localhost
+    {
+        Language Integration_Test
+        {
+            IsSingleInstance = "Yes"
             LocationID = $LocationID
             MUILanguage = $MUILanguage
             MUIFallbackLanguage = $MUIFallbackLanguage
@@ -37,10 +45,8 @@ configuration MSFT_Language_Config {
             AddInputLanguages = $AddInputLanguages
             RemoveInputLanguages = $RemoveInputLanguages
             UserLocale = $UserLocale
-            CopySystem = $true 
+            CopySystem = $true
             CopyNewUser = $true
         }
     }
 }
-
-# TODO: (Optional): Add More Configuration Templates