+++ Tests/Integration/MSFT_LanguagePack.Integration.Tests.ps1
@@ -22,49 +22,49 @@ Import-Module -Name (Join-Path -Path (Join-Path -Path (Split-Path $PSScriptRoot 
 # Using try/finally to always cleanup.
 try
 {
-    $languagePackFolderLocation = "c:\LanguagePacks\"
-    $languagePackFileLocation = "c:\LanguagePacks\x64fre_Server_de-de_lp.cab"
+    $languagePackFolderLocation = 'C:\LanguagePacks\'
+    $languagePackFileLocation = 'C:\LanguagePacks\x64fre_Server_de-de_lp.cab'
     $newLanguagePackFromFolder = 'en-GB'
     $newLanguagePackFromFile = 'de-DE'
     $removeLanguagePack = 'en-US'
+
     if ($env:APPVEYOR -eq $true)
     {
         Write-Warning -Message ('Pre-flight checks for {0} Integration test will be skipped because appveyor does not have the required cab files.' -f $script:DSCResourceName)
     }
     else
     {
-        Describe "Pre-flight Checks" -Tag @('Integration','RequiresDependency') {
-
-            Context "Ensure Language Binaries are available" {
+        Describe 'Pre-flight Checks' -Tag @('Integration','RequiresDependency') {
+            Context 'Ensure Language Binaries are available' {
                 It "Language Pack Folder $languagePackFolderLocation Exists" {
                     Test-Path -Path $languagePackFolderLocation -PathType Container | Should -Be $true
                 }
 
-                It "Language Pack Folder must include at least 1 cab file" {
-                    (Get-ChildItem -Path $languagePackFolderLocation -Filter "*.cab").count | Should -BeGreaterThan 0
+                It 'Language Pack Folder must include at least 1 cab file' {
+                    (Get-ChildItem -Path $languagePackFolderLocation -Filter '*.cab').count | Should -BeGreaterThan 0
                 }
 
-                It "Language Pack File Location must be a cab file" {
+                It 'Language Pack File Location must be a cab file' {
                     $languagePackFileLocation.EndsWith(".cab") | Should -Be $true
                 }
 
-                It "Language Pack File Location must exist" {
+                It 'Language Pack File Location must exist' {
                     Test-Path -Path $languagePackFileLocation -PathType Leaf | Should -Be $true
                 }
             }
 
-            Context "Ensure System requires modification" {
-                It "New Language Pack $newLanguagePackFromFolder mustn't be installed"        {
+            Context 'Ensure System requires modification' {
+                It "New Language Pack $newLanguagePackFromFolder mustn't be installed" {
                     $currentState = Get-TargetResource -LanguagePackName $newLanguagePackFromFolder
                     $currentState.ensure | Should -Be "Absent"
                 }
 
-                It "New Language Pack $newLanguagePackFromFile mustn't be installed"        {
+                It "New Language Pack $newLanguagePackFromFile mustn't be installed" {
                     $currentState = Get-TargetResource -LanguagePackName $newLanguagePackFromFile
                     $currentState.ensure | Should -Be "Absent"
                 }
 
-                It "Language Pack to be removed $removeLanguagePack must be installed"        {
+                It "Language Pack to be removed $removeLanguagePack must be installed" {
                     $currentState = Get-TargetResource -LanguagePackName $removeLanguagePack
                     $currentState.ensure | Should -Be "Present"
                 }
@@ -82,14 +82,14 @@ try
     }
     else
     {
-        Describe "$($script:DSCResourceName) Folder Install Integration" -Tag @("Integration","RequiresDependency") {
+        Describe "$($script:DSCResourceName) Folder Install Integration" -Tag @('Integration','RequiresDependency') {
             #region DEFAULT TESTS
-            It "Should compile and apply the MOF without throwing" {
+            It 'Should compile and apply the MOF without throwing' {
                 {
-                    Write-Verbose "Run Config" -Verbose:$true
-                    & "$($script:DSCResourceName)_Config" -OutputPath $testDrive -LangaugePackName $newLanguagePackFromFolder -LangaugePackLocation $languagePackFolderLocation -Ensure "Present"
+                    Write-Verbose 'Run Config' -Verbose:$true
+                    & "$($script:DSCResourceName)_Config" -OutputPath $testDrive -LangaugePackName $newLanguagePackFromFolder -LangaugePackLocation $languagePackFolderLocation -Ensure 'Present'
                     Start-DscConfiguration -Path $testDrive `
-                        -ComputerName localhost -Wait -Verbose -Force
+                        -ComputerName localhost -Wait -Verbose -Force -ErrorAction Stop
                 } | Should -Not -Throw
             }
 
@@ -111,9 +111,7 @@ try
     }
     else
     {
-        Describe "$($script:DSCResourceName) File Install Integration" -Tag @("Integration","RequiresDependency") {
-
-
+        Describe "$($script:DSCResourceName) File Install Integration" -Tag @('Integration','RequiresDependency') {
             $configMof = (Join-Path -Path $testDrive -ChildPath 'localhost.mof')
 
             It 'Should compile the MOF without throwing' {
@@ -149,7 +147,7 @@ try
     }
     else
     {
-        Describe "$($script:DSCResourceName) Language Pack Uninstall Integration" -Tag @("Integration","RebootRequired") {
+        Describe "$($script:DSCResourceName) Language Pack Uninstall Integration" -Tag @('Integration','RebootRequired') {
             It 'Should compile the MOF without throwing' {
                 {
                     & "$($script:DSCResourceName)_Config" `
@@ -184,6 +182,4 @@ finally
     Restore-TestEnvironment -TestEnvironment $testEnvironment
 
     #endregion
-
-    # TODO: Other Optional Cleanup Code Goes Here...
 }