+++ Tests/Unit/MSFT_ScheduledTask.Tests.ps1
@@ -104,7 +104,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Minutes 150).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith { return $null }
@@ -132,7 +132,7 @@ try
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Minutes 15).ToString()
                     Ensure             = 'Absent'
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -176,7 +176,7 @@ try
                     TaskName = 'Test task'
                     TaskPath = '\Test\'
                     Enable   = $false
-                    Verbose  = $True
+                    Verbose  = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -221,7 +221,7 @@ try
                     TaskName = 'Test task'
                     TaskPath = '\Test\'
                     Ensure   = 'Absent'
-                    Verbose  = $True
+                    Verbose  = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -273,7 +273,7 @@ try
                     ActionExecutable = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
                     ScheduleType     = 'Once'
                     Ensure           = 'Absent'
-                    Verbose          = $True
+                    Verbose          = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask
@@ -296,7 +296,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Minutes 150).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -348,7 +348,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Minutes 30).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -395,7 +395,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Hours 4).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -447,7 +447,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Hours 4).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -493,7 +493,7 @@ try
                     ActionExecutable = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
                     ScheduleType     = 'Daily'
                     DaysInterval     = 3
-                    Verbose          = $True
+                    Verbose          = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -544,7 +544,7 @@ try
                     ActionExecutable = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
                     ScheduleType     = 'Daily'
                     DaysInterval     = 3
-                    Verbose          = $True
+                    Verbose          = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -589,7 +589,7 @@ try
                     RepeatInterval      = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration  = (New-TimeSpan -Hours 8).ToString()
                     ExecuteAsCredential = New-Object System.Management.Automation.PSCredential ('DEMO\RightUser', (ConvertTo-SecureString 'ExamplePassword' -AsPlainText -Force))
-                    Verbose             = $True
+                    Verbose             = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -643,7 +643,7 @@ try
                     RepetitionDuration  = (New-TimeSpan -Hours 8).ToString()
                     ExecuteAsCredential = New-Object System.Management.Automation.PSCredential ('DEMO\RightUser', (ConvertTo-SecureString 'ExamplePassword' -AsPlainText -Force))
                     LogonType           = 'S4U'
-                    Verbose             = $True
+                    Verbose             = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -699,7 +699,7 @@ try
                     RepetitionDuration  = (New-TimeSpan -Hours 8).ToString()
                     ExecuteAsCredential = New-Object System.Management.Automation.PSCredential ('DEMO\RightUser', (ConvertTo-SecureString 'ExamplePassword' -AsPlainText -Force))
                     RunLevel            = 'Highest'
-                    Verbose             = $True
+                    Verbose             = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -754,7 +754,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -808,7 +808,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -862,7 +862,7 @@ try
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
                     Enable             = $false
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -921,7 +921,7 @@ try
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
                     ExecutionTimeLimit = (New-TimeSpan -Seconds 0).ToString()
                     Enable             = $true
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -984,7 +984,7 @@ try
                     ExecutionTimeLimit = (New-TimeSpan -Minutes 7).ToString()
                     RestartInterval    = (New-TimeSpan -Minutes 8).ToString()
                     Enable             = $true
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1043,7 +1043,7 @@ try
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
                     Enable             = $false
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1095,7 +1095,7 @@ try
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
                     Enable             = $true
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1151,7 +1151,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 15).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 8).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1229,7 +1229,7 @@ try
                     IdleDuration       = (New-TimeSpan -Minutes 6).ToString()
                     ExecutionTimeLimit = (New-TimeSpan -Minutes 7).ToString()
                     RestartInterval    = (New-TimeSpan -Minutes 8).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1295,7 +1295,7 @@ try
                     IdleDuration       = (New-TimeSpan -Minutes 6).ToString()
                     ExecutionTimeLimit = (New-TimeSpan -Minutes 7).ToString()
                     RestartInterval    = (New-TimeSpan -Minutes 8).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1357,7 +1357,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 20).ToString()
                     RepetitionDuration = 'Indefinitely'
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1410,7 +1410,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 20).ToString()
                     RepetitionDuration = (New-TimeSpan -Hours 9).ToString()
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1463,7 +1463,7 @@ try
                     ScheduleType       = 'Once'
                     RepeatInterval     = (New-TimeSpan -Minutes 20).ToString()
                     RepetitionDuration = 'Indefinitely'
-                    Verbose            = $True
+                    Verbose            = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1508,7 +1508,7 @@ try
                     TaskName = 'Test task'
                     TaskPath = '\Test\'
                     Enable   = $false
-                    Verbose  = $True
+                    Verbose  = $true
                 }
 
                 Mock -CommandName Get-ScheduledTask -MockWith {
@@ -1524,13 +1524,14 @@ try
                                 Interval = "PT15M"
                             }
                             CimClass   = @{
-                                CimClassName = 'MSFT_TaskEventTrigger'
+                                CimClassName = 'MSFT_TaskSessionStateChangeTrigger'
                             }
                         }
                         Settings = [pscustomobject] @{
                             Enabled = $true
                         }
-                    } }
+                    } 
+                }
 
                 It 'Should return the correct values from Get-TargetResource' {
                     $result = Get-TargetResource @testParameters
@@ -1546,6 +1547,187 @@ try
                 It 'Should disable the scheduled task in the set method' {
                     Set-TargetResource @testParameters
                     Assert-MockCalled Register-ScheduledTask -Exactly -Times 1
+                }
+            }
+
+            Context 'When a scheduled task with an OnEvent scheduletype is in desired state' {
+                $testParameters = @{
+                    TaskName          = 'Test task'
+                    TaskPath          = '\Test\'
+                    ScheduleType      = 'OnEvent'
+                    ActionExecutable  = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
+                    EventSubscription = '<QueryList><Query Id="0" Path="System"><Select Path="System">*[System[Provider[@Name=''User32''] and EventID=1600]]</Select></Query></QueryList>'
+                    Delay             = '00:01:00'
+                    Enable            = $true
+                    Verbose           = $true
+                }
+
+                Mock -CommandName Get-ScheduledTask -MockWith {
+                    @{
+                        TaskName = $testParameters.TaskName
+                        TaskPath = $testParameters.TaskPath
+                        Actions  = [pscustomobject] @{
+                            Execute = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
+                        }
+                        Triggers = [pscustomobject] @{
+                            Delay = 'PT1M'
+                            Subscription = $testParameters.EventSubscription
+                            CimClass   = @{
+                                CimClassName = 'MSFT_TaskEventTrigger'
+                            }
+                        }
+                        Settings = [pscustomobject] @{
+                            Enabled = $true
+                        }
+                    } 
+                }
+
+                It 'Should return the correct values from Get-TargetResource' {
+                    $result = Get-TargetResource @testParameters
+                    $result.Enable | Should -Be $true
+                    $result.Ensure | Should -Be 'Present'
+                    $result.ScheduleType | Should -Be 'OnEvent'
+                    $result.EventSubscription | Should -Be $testParameters.EventSubscription
+                    $result.Delay | Should -Be $testParameters.Delay
+                }
+
+                It 'Should return true from the test method' {
+                    Test-TargetResource @testParameters | Should -Be $true
+                }
+            }
+
+            Context 'When a scheduled task with an OnEvent scheduletype needs to be created' {
+                $testParameters = @{
+                    TaskName          = 'Test task'
+                    TaskPath          = '\Test\'
+                    ScheduleType      = 'OnEvent'
+                    ActionExecutable  = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
+                    EventSubscription = '<QueryList><Query Id="0" Path="System"><Select Path="System">*[System[Provider[@Name=''User32''] and EventID=1600]]</Select></Query></QueryList>'
+                    Delay             = '00:01:00'
+                    Enable            = $true
+                    Verbose           = $true
+                }
+
+                Mock -CommandName Get-ScheduledTask
+
+                It 'Should return the correct values from Get-TargetResource' {
+                    $result = Get-TargetResource @testParameters
+                    $result.Ensure | Should -Be 'Absent'
+                }
+
+                It 'Should return false from the test method' {
+                    Test-TargetResource @testParameters | Should -Be $false
+                }
+
+                It 'Should register the new scheduled task' {
+                    Set-TargetResource @testParameters
+                    Assert-MockCalled Register-ScheduledTask -Exactly -Times 1 -Scope It
+                }
+            }
+
+            Context 'When a scheduled task with an OnEvent scheduletype needs to be updated' {
+                $testParameters = @{
+                    TaskName          = 'Test task'
+                    TaskPath          = '\Test\'
+                    ScheduleType      = 'OnEvent'
+                    ActionExecutable  = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
+                    EventSubscription = '<QueryList><Query Id="0" Path="System"><Select Path="System">*[System[Provider[@Name=''User32''] and EventID=1600]]</Select></Query></QueryList>'
+                    Delay             = '00:05:00'
+                    Enable            = $true
+                    Verbose           = $true
+                }
+
+                Mock -CommandName Get-ScheduledTask -MockWith {
+                    @{
+                        TaskName = $testParameters.TaskName
+                        TaskPath = $testParameters.TaskPath
+                        Actions  = [pscustomobject] @{
+                            Execute = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
+                        }
+                        Triggers = [pscustomobject] @{
+                            Delay = 'PT1M'
+                            Subscription = '<QueryList><Query Id="0" Path="System"><Select Path="System">*[System[Provider[@Name=''User32''] and EventID=1601]]</Select></Query></QueryList>'
+                            CimClass   = @{
+                                CimClassName = 'MSFT_TaskEventTrigger'
+                            }
+                        }
+                        Settings = [pscustomobject] @{
+                            Enabled = $true
+                        }
+                    } 
+                }
+
+                It 'Should return the correct values from Get-TargetResource' {
+                    $result = Get-TargetResource @testParameters
+                    $result.Enable | Should -Be $true
+                    $result.Ensure | Should -Be 'Present'
+                    $result.ScheduleType | Should -Be 'OnEvent'
+                }
+
+                It 'Should return false from the test method' {
+                    Test-TargetResource @testParameters | Should -Be $false
+                }
+
+                It 'Should not call Register-ScheduledTask on an already registered task' {
+                    Set-TargetResource @testParameters
+                    Assert-MockCalled Register-ScheduledTask -Times 0 -Scope It
+                }
+
+                It 'Should call Set-ScheduledTask to update the scheduled task with the new values' {
+                    Set-TargetResource @testParameters
+                    Assert-MockCalled Set-ScheduledTask -Times 1 -Scope It
+                }
+            }
+
+            Context 'When a scheduled task with an OnEvent scheduletype is used on combination with unsupported parameters for this scheduletype' {
+                $testParameters = @{
+                    TaskName          = 'Test task'
+                    TaskPath          = '\Test\'
+                    ScheduleType      = 'OnEvent'
+                    ActionExecutable  = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
+                    EventSubscription = '<QueryList><Query Id="0" Path="System"><Select Path="System">*[System[Provider[@Name=''User32''] and EventID=1600]]</Select></Query></QueryList>'
+                    RandomDelay       = '01:00:00'
+                    Delay             = '00:01:00'
+                    Enable            = $true
+                    Verbose           = $true
+                }
+
+                Mock -CommandName Get-ScheduledTask -MockWith {
+                    @{
+                        TaskName = $testParameters.TaskName
+                        TaskPath = $testParameters.TaskPath
+                        Actions  = [pscustomobject] @{
+                            Execute = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
+                        }
+                        Triggers = [pscustomobject] @{
+                            Delay = 'PT1M'
+                            Subscription = $testParameters.EventSubscription
+                            CimClass   = @{
+                                CimClassName = 'MSFT_TaskEventTrigger'
+                            }
+                        }
+                        Settings = [pscustomobject] @{
+                            Enabled = $true
+                        }
+                    } 
+                }
+
+                It 'Should return the correct values from Get-TargetResource' {
+                    $result = Get-TargetResource @testParameters
+                    $result.Enable | Should -Be $true
+                    $result.Ensure | Should -Be 'Present'
+                    $result.ScheduleType | Should -Be 'OnEvent'
+                    $result.RandomDelay | Should -Be '00:00:00'
+                }
+
+                It 'Should return true from the test method - ignoring the RandomDelay parameter' {
+                    Test-TargetResource @testParameters | Should -Be $true
+                }
+
+                $testParameters.EventSubscription = 'InvalidXML'
+
+                It 'When an EventSubscription cannot be parsed as valid XML an error is generated when changing the task' {
+                    { Set-TargetResource @testParameters } | Should throw
                 }
             }
         }